The pretty printer gets confused by comments including unbalanced single or double quotes, like in this interface file:
-------------
%typemap(perl5, check) size_t *SIZE_T_VECTORINOUT
{
    /* NEXT LINE AND REST OF WRAPPER WILL BE MISSING. */
    /* Do nothing (don't check for NULL pointer) */
}

void x(size_t *SIZE_T_VECTORINOUT);
-------------

The code tries to find the end of the character/string constant. When
it reaches EOF, nothing will get emitted because no \n is seen.

This is a change to emit the last line of a string. But the pretty
printer probably should be fixed to handle comments.