All but two test-suite tests succeed:
funcptr fails: the type is bunched in front of the variable name, so
that it doesn't parse right in C.  The type of the function pointer
needs to wrap.  I will ask Mr. Beazley how this is done.

template_ns4: I don't even understand what this code does.  I need
to ask for help on it.