Fixed template-integer argument bug.  This is kind of wacky.  Basically, if you have this:
    const int blah = 4;

and you say this:

    %template(whatever) sometemplate<blah>;

it gets resolved *in the parser* to

    %template(whatever) sometemplate<4>;

Not only that, this applies to *ALL* occurences of the template parameters. For
example, if you had this:

    void foo(sometemplate<blah> x);

it really gets parsed as this:

    void foo(sometemplate<4> x);

This means that "sometemplate<blah>" will never actually appear in the generated
wrapper code. Instead, "sometemplate<4>" is used.

I'm not sure if this is the best solution, but it is the easiest
solution and it passes the Marcelo evil test. Otherwise, I have to
make a bunch of hellish modifications to the type system to handle
integer value resolution in template type arguments (bleah!).