Fixed the behaviour of static const members, now in the following case:
     struct CavityPackFlags
     {
       typedef unsigned int viewflags;
       static const viewflags forward_field  = 1 << 0;
       static const viewflags cavity_flags;
     };

forward_field and  cavity_flags are treated different. In
the first case, a constant (like enum) is emmited,
in the second case, a const variable is added to cvar.

This is necessary because gcc3.1.1
doesn't even generate a reference for forward_field,
and therefore it complains later about an undefined symbol
if you try to add it to the cvar entity, because
a reference for forward_field is needed in that case.

Also, thanks to the C++ restrictions, there is no way you
can have a static const member variable with initial value
which is not really really a constant. And the
types allowed in such declarations are also very limited (int, unsigned int,
etc), so, treat them as %constant plays well with the
target language too.

Note that in C++ forward_field and cavity_flags are really
very different animals, in fact, if you have

template<CavityFlags::viewflags>
struct Hello {
};

then, the compiler allows you to do

   Hello<CavityFlags::forward_field> h;

but this is forbidden

   Hello<CavityFlags::cavity_flags> h;


and well, this makes sense

    unsigned int *ptr = &CavityFlags::cavity_flags;

and this doesn't

    unsigned int *ptr = &CavityFlags::forward_field;

but this last problem maybe only detected at linking or running time.