Add a check for PCRE library to configure.
We use standard AX_PATH_GENERIC() macro from Autoconf archive to do the check
to keep our own code as simple as possible. Notice that this script must be
used, just adding the include and library directories to {C,LD}FLAGS is not
enough, notably -DPCRE_STATIC which is output by pcre-config for static builds
only is crucial. Also use LIBS instead of LDFLAGS to fix linking when using
static libraries.

Also note that this allows to pass PCRE_CONFIG variable value to configure to
force the use of the specified script (and not the one first found in PATH),
which is especially important when cross-compiling.

Finally, PCRE is required by default now, --without-pcre must be explicitly
used to build without it.