Fix the rename_pcre_encoder unit test for Python.
Due to a bug in the way %rename() directives without a name currently work,
the second %rename in rename_pcre_encoder.i overrides the first one and the
function wxDoSomething() is not renamed to DoSomething() any more. Remove the
test for this renaming that doesn't work to at least let the rest of the test
to pass.

Also test that SetXXX() functions from the .i file really are renamed to
putXXX() as expected.